LPS-108118 Move to the floating toolbar folder
