+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/RowWithControls.js
@@ -39,7 +39,7 @@ import selectShowLayoutItemTopper from '../../selectors/selectShowLayoutItemTopp
 import {useDispatch, useSelector} from '../../store/index';
 import duplicateItem from '../../thunks/duplicateItem';
 import resizeColumns from '../../thunks/resizeColumns';
-import SaveFragmentCompositionModal from '../SaveFragmentSaveFragmentCompositionModal';
+import SaveFragmentCompositionModal from '../SaveFragmentCompositionModal';
 import Topper from '../Topper';
 import FloatingToolbar from '../floating-toolbar/FloatingToolbar';
 import ColumnOverlayGrid from './ColumnOverlayGrid';
