+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/SaveFragmentCompositionModal.js
@@ -22,9 +22,9 @@ import ClaySticker from '@clayui/sticker';
 import PropTypes from 'prop-types';
 import React, {useContext, useState} from 'react';
 
-import Button from '../../common/components/Button';
-import {openImageSelector} from '../../core/openImageSelector';
-import {ConfigContext} from '../config/index';
+import Button from '../../../common/components/Button';
+import {openImageSelector} from '../../../core/openImageSelector';
+import {ConfigContext} from '../../config/index';
 
 const SaveFragmentCompositionModal = ({
 	errorMessage,
